add to cart

+++ app/src/main/java/com/mad43/stylista/data/remote/entity/auth/Customer.kt
@@ -6,7 +6,7 @@ data class Customer(
     val first_name: String,
     val id: Long,
     val last_name: String,
-    val last_order_id: Int,
+    val last_order_id: Long,
     val last_order_name: String,
     val marketing_opt_in_level: Any,
     val multipass_identifier: Any,
